backend/requirements.txt
# FreshKeep Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database - Supabase
supabase>=2.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.24.0
aiohttp>=3.9.0

# Background Tasks
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# File Handling
aiofiles>=23.2.0
Pillow>=10.4.0

# Caching (optional)
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
